Fighting MS VC compiler warnings

+++ src/fontio/logic/cff/CffReader.hpp
@@ -321,7 +321,7 @@ namespace fontio { namespace logic { namespace cff
 
             if (format == 0)
             {
-                for (size_t i = 1; i < totalGlyphs; i++)
+                for (uint16_t i = 1; i < totalGlyphs; i++)
                 {
                     gidToSid[i] = this->ReadBigEndian<uint16_t>(stream);
                 }
